Fix: Crash in recent nvim versions
After not updating nvim for a few weeks, recent versions made some
changes (good ones!) that revealed an error in the Finder:open()
function.

The prompt BufLeave autocommand was registered before the results window
was created. When creating the results window, the prompt buffer and
window were closed and removed, and this caused an abrupt crash of nvim.

The solution is to wait until the prompt window is focused and already
in insert mode before registering keybinds and autocommands.