Finalized search finder (#4)
This last commit ensures that the quickfix list is always filled no
matter the action taken from the search finder.

This completes the basic implementation of the search finder. This
required a lot of changes to finder.lua, so some refactoring is needed
to clean up and ensure a tidy api for finders in general.

After that, a release for v0.1 would make sense!

Closes #4