Add cache and fix errors
There were some issues when adding the cache that are now handled. This
was due to iterators returning nil to signal completion, but I wanted to
use nil to represent a completed iterator. In that case, a bit more code
was needed to properly signal a stopped or completed iterator.