Updated filename filter
Renamed from "file" to "filename" filter. This filter attempts to be a
very simple yet usable file finder. Rather than fuzzy matching and
returning many results, it narrows down the results by prioritizing
matches on the filename.

The query is broken into tokens separated by whitespace. The first token
is used to filter the matches on the filename. If there is a match in
the filename then any remaining tokens are matched against the entire
path. If all match, then the path is kept, otherwise it is discarded.

There are still some potential improvements including:
* Ignoring delimiter characters like -_.
* Ignoring case