Filters: Add wrap filter
The purpose of the wrap filter is to take each line and wrap it in a
table. By default the only entry is result, but a custom function can be
passed in for special cases like ripgrep.

Perhaps this could be used to also split on new lines to reduce that
logic?