Add transient option to finders
Transient finders are those that are closed quickly with esc. This is
needed because some finders benefit from being open longer, like search
with ripgrep. But some finders are used frequently enough (files,
buffers) that it might be annoying to always require <c-c> or esc twice
to close, so that is why transient is needed.

Transient defaults to false.