Add ignore paths config to filename filter
Adds a new config.ignore list that acts as an ignore list for the
filepath filter. All strings in the table are prefixed with "^" and
suffixed with "$" to match the entire string. To match partial strings
use "*" at any point.

For example:
- to match all .png files use "*.png"
- to match any file containing the word backup use "*backup*"

This helps to ignore certain directories like __pycache__ or filetypes
that are part of repos but aren't needed for edits in nvim.